requirements.txt
# Kotori.ai - Requirements for Deployment
# Core web framework
streamlit>=1.28.0

# AI/ML Libraries
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0

# Vector Database
chromadb>=0.4.0

# Embeddings and NLP
sentence-transformers>=2.2.0
huggingface-hub>=0.17.0

# Environment and Configuration
python-dotenv>=1.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# HTTP Requests (if needed for external APIs)
requests>=2.31.0

# Path handling (usually included with Python 3.4+)
pathlib2>=2.3.7; python_version < "3.4"

# Optional: For better performance
faiss-cpu>=1.7.4

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
